% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_corr.R
\name{cyt_corr}
\alias{cyt_corr}
\title{Calculate correlations between a target variable and all other numeric variables.}
\usage{
cyt_corr(
  data,
  target,
  methods = c("spearman", "pearson"),
  group_var = NULL,
  compare_groups = FALSE,
  plot = FALSE,
  progress = NULL
)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{target}{A character string specifying the name of the target variable.
Must be a numeric column in \code{data}.}

\item{methods}{A character string indicating the correlation method to be used.
Can be "pearson" (default) or "spearman".}

\item{group_var}{An optional character string specifying the name of a grouping
variable in \code{data}. If provided, group-wise correlations will be calculated.}

\item{compare_groups}{Logical. If \code{TRUE} and \code{group_var} is provided with at
least two levels, the function will compare the correlations between the
first two levels of \code{group_var}.}

\item{plot}{Logical. If \code{TRUE}, the function will generate and return a correlation plot.}
}
\value{
A list containing:
\item{results}{A data frame with overall correlation results, including
variable names, correlation coefficients (\code{r}), p-values (\code{p}), sample
size (\code{n}), method, and adjusted p-values (\code{p_bonf}, \code{p_bh}), sorted by
absolute correlation coefficient in descending order.}
\item{results_groupwise}{A data frame with group-wise correlation results,
if \code{group_var} is provided. Includes similar columns as \code{results}, plus
the \code{group} identifier.}
\item{results_diff}{A data frame with results of correlation comparison
between the first two groups, if \code{compare_groups} is \code{TRUE} and
\code{group_var} has at least two levels. Includes \code{variable}, \code{r_diff}
(difference in correlations), \code{z} (Fisher's z-score), \code{p_diff} (p-value
for the difference), and adjusted p-values (\code{p_diff_bonf}, \code{p_diff_bh}).}
\item{plot}{A correlation plot, if \code{plot} is \code{TRUE}.}
}
\description{
This function computes Pearson or Spearman correlation coefficients between a
specified target variable and all other numeric variables in the input data frame.
It also calculates p-values and adjusted p-values (Bonferroni and Benjamini-Hochberg).
Optionally, it can perform group-wise correlation analysis and compare correlations
between two groups using Fisher's z-transformation.
}
