% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_mint_splsda.R
\name{cyt_mint_splsda}
\alias{cyt_mint_splsda}
\title{Analyze data with MINT Sparse Partial Least Squares Discriminant Analysis (sPLS-DA).}
\usage{
cyt_mint_splsda(
  data,
  group_col,
  batch_col,
  group_col2 = NULL,
  colors = NULL,
  output_file = NULL,
  ellipse = TRUE,
  bg = FALSE,
  var_num = 20,
  comp_num = 2,
  scale = NULL,
  pch_values = NULL,
  roc = FALSE,
  progress = NULL
)
}
\arguments{
\item{data}{A matrix or data frame containing the variables. Columns not
specified by \code{group_col}, \code{group_col2}, or \code{multilevel_col} are assumed to be continuous
variables for analysis.}

\item{group_col}{A string specifying the first grouping column name that contains grouping
information. If \code{group_col2} is not provided, it will be used for both
grouping and treatment.}

\item{batch_col}{A string specifying the batch column name that contains batch
information.}

\item{group_col2}{A string specifying the second grouping column name. Default is
\code{NULL}.}

\item{colors}{A vector of colors for the groups or treatments. If
\code{NULL}, a random palette (using \code{rainbow}) is generated based on
the number of groups.}

\item{output_file}{A string specifying the file name for saving the PDF output.
If set to NULL, the function runs in interactive mode.}

\item{ellipse}{Logical. Whether to draw a 95\\% confidence ellipse on the figures.
Default is \code{FALSE}.}

\item{bg}{Logical. Whether to draw the prediction background in the figures.
Default is \code{FALSE}.}

\item{var_num}{Numeric. The number of variables to be used in the PLS-DA model.}

\item{comp_num}{Numeric. The number of components to calculate in the sPLS-DA model.
Default is 2.}

\item{scale}{Character. Option for data transformation; if set to \code{"log2"}, a log2
transformation is applied to the continuous variables. Default is \code{NULL}.}

\item{pch_values}{A vector of integers specifying the plotting characters (pch values)
to be used in the plots.}

\item{roc}{Logical. Whether to compute and plot the ROC curve for the model.
Default is \code{FALSE}.}
}
\value{
In Download mode, a PDF file is written. In Interactive mode, a named list
(\code{results_list}) of plots and results is returned. If \code{group_col2} is used,
a nested list is returned, with each element corresponding to a level of \code{group_col2}.
}
\description{
This function performs a MINT (Multivariate INTegrative) sPLS-DA to handle
batch effects by modeling a global biological signal across different studies or batches.
If a second grouping column (\code{group_col2}) is provided, the analysis is stratified
and performed for each level of that column.
}
\examples{
# Loading ExampleData5 dataset with batch column
data_df <- ExampleData5[,-c(2,4)]
data_df <- dplyr::filter(data_df, Group != "ND")

cyt_mint_splsda(data_df, group_col = "Group",
 batch_col = "Batch", colors = c("black", "purple"),
 ellipse = TRUE, var_num = 25, comp_num = 2,
 scale = "log2")
}
