% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData3}
\alias{ExampleData3}
\title{Example Cytokine Profiling Data 3.}
\format{
A data frame with 64 rows and 14 columns:
\describe{
\item{Stimulation}{Stimulation assigned to the subjects.}
\item{Group}{Group assigned to the subjects.}
\item{...}{Additional numeric columns representing measured cytokines.}
}
@source Example data compiled for cytokine profiling.
@references
SantaCruz-Calvo S, Saraswat S, Hasturk H, Dawson DR, Zhang XD,
Nikolajczyk BS. Periodontitis and Diabetes Differentially Affect
Inflammation in Obesity. \emph{J Dent Res.} 2024;103(12):1313-1322.
doi:10.1177/00220345241280743

@examples
data(ExampleData3)
}
\usage{
ExampleData3
}
\description{
Contains observed values of cytokines and their respective treatment
and groups, derived from:
}
\keyword{datasets}
