% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyt_heatmap.R
\name{cyt_heatmap}
\alias{cyt_heatmap}
\title{Heat Map.}
\usage{
cyt_heatmap(
  data,
  scale = c(NULL, "log2", "row_zscore", "col_zscore"),
  annotation_col = NULL,
  annotation_side = c("auto", "row", "col"),
  progress = NULL
)
}
\arguments{
\item{data}{A data frame containing the input data. Only numeric columns will be used.}

\item{scale}{Character. Optional scaling method. One of NULL (no scaling), "log2" (log2 transformation), "row_zscore" (z-score normalization by row), or "col_zscore" (z-score normalization by column). Default is NULL.}

\item{annotation_col}{Optional. A character string specifying a column name in \code{data} or a vector of annotations matching the number of rows or columns in \code{data} for adding colored annotations to the heatmap.}

\item{annotation_side}{Character. Specifies whether the annotation should be applied to rows or columns. One of "auto" (automatically determine based on length), "row", or "col". Default is "auto".}
}
\value{
A pheatmap object representing the generated heatmap.
}
\description{
Heat Map.
}
\examples{
# Load sample data
data("ExampleData1")
data_df <- ExampleData1
# Generate a heatmap with log2 scaling and annotation based on
# the "Group" column
cyt_heatmap(
  data = data_df[, -c(2:3)],
  scale = "log2",  # Optional scaling
  annotation_col_name = "Group"
)

}
